Move common folder in UI components to UI folder
