+++ client/app/ui/AgreementModal.tsx
