--- client/app/ui/common/index.ts
@@ -1,10 +1,0 @@
-export * from './dnd-kit'
-
-export * from './AgreementModal'
-export * from './Button'
-export * from './Container'
-export * from './CheckCircle'
-export * from './Divider'
-export * from './Skeleton'
-export * from './Title'
-export * from './Input'
