+++ client/app/ui/dnd-kit/SortableOverlay.tsx
